Refactoriza funciones
