+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import "./globals.css"
+import "../styles/globals.css"
 
 export default function RootLayout({
   children
