--- src/lib/classnames.ts
@@ -1,5 +1,0 @@
-export default function classNames(
-  ...classes: (string | undefined | null | false)[]
-) {
-  return classes.filter(Boolean).join(" ")
-}
