--- src/lib/getLevel.ts
@@ -1,4 +1,0 @@
-export default function getLevel(levelPoints: number | undefined): number {
-  const level = !levelPoints ? 1 : Math.floor(levelPoints / 1000) + 1
-  return level
-}
