+++ src/lib/next-auth.d.ts
